feat(balance): Filter on whole year for reimbursements
