+++ src/ducks/reimbursements/ReimbursementsPage.jsx
@@ -1,16 +1,14 @@
 import React from 'react'
 import BarTheme from 'ducks/mobile/BarTheme'
+import { connect } from 'react-redux'
+import { flowRight as compose } from 'lodash'
+import { translate, withBreakpoints } from 'cozy-ui/react'
 import Header from 'components/Header'
 import { Padded } from 'components/Spacing'
 import { PageTitle } from 'components/Title'
 import cx from 'classnames'
-import { translate, withBreakpoints } from 'cozy-ui/react'
-import { flowRight as compose } from 'lodash'
 import BackButton from 'components/BackButton'
 import { ConnectedSelectDates } from 'components/SelectDates'
-import { queryConnect } from 'cozy-client'
-import { transactionsConn } from 'doctypes'
-import { connect } from 'react-redux'
 import HealthReimbursements from 'ducks/reimbursements/HealthReimbursements'
 
 function getSubComponent(filteringDoc) {
@@ -64,9 +62,6 @@ function mapStateToProps(state) {
 const ReimbursementsPage = compose(
   translate(),
   withBreakpoints(),
-  queryConnect({
-    transactions: transactionsConn
-  }),
   connect(mapStateToProps)
 )(RawReimbursementsPage)
 export default ReimbursementsPage
